Mas cambios de variables

+++ libawui/awui/Emulation/MasterSystem/VDP.h
@@ -65,7 +65,7 @@ namespace awui {
 
 					uint8_t GetSpriteColor(uint16_t sprite, int x, int y, bool flipx, bool flipy, bool otherPalete) const;
 					bool GetSpritePixel(uint8_t x, uint8_t y, uint8_t * color) const;
-					uint8_t GetBackgroundPixel(uint16_t sprite, int x, int y, bool flipx, bool flipy, bool otherPalete) const;
+					uint8_t GetBackgroundPixel(uint16_t sprite, int16_t x, int16_t y, bool flipx, bool flipy, bool otherPalete) const;
 
 				public:
 					VDP(CPU * cpu);
